+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  * Default bundler version is now 1.1.3.
+  * Default bundler version is now 1.1.5.
   * Deploy hooks now have access to `account_name` and `environment_name`.
   * Improves deploy output, especially for `--verbose`.
   * Sends all log output through a new Shell object that formats and adds timestamps.